Update to 2.0.0.
Add more info about mcd in the readme.
Change pusdh to the cd function in the readme.
Add an OPTIONS... to mcd
Add the shell word in the Description.
Add cat to the ExtComm.
Add "${____D_i_r_S_t_a_c_k____s[@]}" array to replace the builtin dirs.
Add a cd function
 * Replace the embedded newline in a directory with literal \n.
 * It Prints the git status if git is installed.
Declare the variables local in mcd and cd.
Check if the tput utility is installed, if it is print bold/colored fonts.
Add a while loop that checks for a command arguments, if any.
Add an about menu and usage menu as an option.