Update to version 2.0.1
Change the variable status to return_status.
Change the variable "${#directory}" to "${#directory[@]}".
Declare the other arrays with in the mcd function.
Change the function __Menu__ to ___mcd__m__e__n__u___
Change the Location to Directory when the Quit is selected in the menu.
Add more info in the about and the Usage function.