Update to version 2.0.6.
Change the visited to remembered in the readme.
Add a [--clear|-c] function to clear the directory stack.
Change the visited to remembered in the mcd script.
Add an option to print the "$OLDPWD" if the - (dash) option is used in cd, (builtin cd behaviour).
Add a __mcd__c__h__e__c__k__g__i__t function.