Update to version 2.0.8.
Add a requirements in the readme.
 * Bash version 4 is a must.
 * Tput and git are optional.

Change the "source ./mcd" to "cd mcd/ && source ./mcd"
Download the archive and unpack it as another option.
Add a file string in the readme.
Add an option to check the bash version, bash version below 4 will exit the mcd function.
Modify the meaning of [-n|--no-color] in the Usage function.